fix(sonnet): 종합 린트 및 코드 품질 개선
- 49개 파일의 코드 품질 이슈 해결
- devcontainer.json JSON 유효성 문제 수정
- 전체 프록시 핸들러 코드 정리 및 최적화
- 서비스 레이어 및 설정 모듈 개선
- 메트릭스 및 미들웨어 코드 품질 향상

+++ internal/services/adapters/upstream_client.go
@@ -102,7 +102,7 @@ func (c *HTTPUpstreamClient) Fetch(ctx context.Context, url string,
 			bytesReceived = resp.ContentLength
 		}
 
-		c.performanceMonitor.RecordRequest(c.proxyType, success, responseTime, bytesReceived, int64(req.ContentLength))
+		c.performanceMonitor.RecordRequest(c.proxyType, success, responseTime, bytesReceived, req.ContentLength)
 
 		if err != nil {
 			c.performanceMonitor.RecordTimeout(c.proxyType)
